@@ -1,17 +1,5 @@
 package life.genny.bridge.live.data;
 
-import java.util.UUID;
-
-import javax.enterprise.event.Observes;
-import javax.inject.Inject;
-import javax.inject.Singleton;
-import javax.json.bind.Jsonb;
-import javax.json.bind.JsonbBuilder;
-
-import org.apache.commons.lang3.StringUtils;
-import org.eclipse.microprofile.config.inject.ConfigProperty;
-import org.jboss.logging.Logger;
-
 import io.quarkus.runtime.StartupEvent;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.bridge.BridgeEventType;
@@ -26,6 +14,16 @@ import life.genny.qwandaq.utils.CommonUtils;
 import life.genny.qwandaq.utils.HttpUtils;
 import life.genny.qwandaq.utils.KafkaUtils;
 import life.genny.serviceq.Service;
+import org.apache.commons.lang3.StringUtils;
+import org.eclipse.microprofile.config.inject.ConfigProperty;
+import org.jboss.logging.Logger;
+
+import javax.enterprise.event.Observes;
+import javax.inject.Inject;
+import javax.inject.Singleton;
+import javax.json.bind.Jsonb;
+import javax.json.bind.JsonbBuilder;
+import java.util.UUID;
 
 /**
  * ExternalConsumer --- External clients can connect to the endpoint configured in {@link
@@ -183,9 +181,8 @@ public class ExternalConsumer {
 	 */
 	void routeDataByMessageType(JsonObject body, GennyToken gennyToken) {
 		// JsonObject nonTokenBody = (JsonObject) body.remove("token");
-		
 
-		log.info("Incoming Payload = " + body.toString());
+		log.info("Incoming Payload = %s".formatted(body));
 
 		if (body == null || body.getString("msg_type") == null) {
 			log.error("Bad body JsonObject passed");
