@@ -21,9 +21,8 @@
 
 package life.genny.qwandaq.attribute;
 
-import java.io.Serializable;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
+import com.querydsl.core.annotations.QueryExclude;
+import life.genny.qwandaq.datatype.DataType;
 
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
@@ -32,10 +31,8 @@ import javax.persistence.InheritanceType;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlRootElement;
-
-import life.genny.qwandaq.datatype.DataType;
-
-import com.querydsl.core.annotations.QueryExclude;
+import java.io.Serializable;
+import java.time.LocalDate;
 
 /**
  * AttributeDate class handles LocalDate based attributes.
@@ -46,13 +43,12 @@ import com.querydsl.core.annotations.QueryExclude;
  * <p>
  * AttributeDate represent the major way of specifying the Date data type about a target
  * from sources.
- * <p>
- * 
- * 
- * @author      Adam Crow
- * @author      Byron Aguirre
- * @version     %I%, %G%
- * @since       1.0
+ * </p>
+ *
+ * @author Adam Crow
+ * @author Byron Aguirre
+ * @version %I%, %G%
+ * @since 1.0
  */
 
 @SuppressWarnings("serial")
@@ -65,27 +61,25 @@ import com.querydsl.core.annotations.QueryExclude;
 @DiscriminatorValue("date")
 public class AttributeDate extends Attribute implements Serializable {
 
-	private static final long serialVersionUID = 1L;
-	
-	/**
-	 * Constructor.
-	 */
-	@SuppressWarnings("unused")
-	public AttributeDate()
-	{
-		super();
-		// dummy for hibernate
-	}
-	
-	/**
-	 * Constructor.
-	 * 
-	 * @param aCode The unique code for this Question
-	 * @param aName The human readable summary name
-	 */
-	public AttributeDate(String aCode, String aName)
-	{
-		super(aCode, aName, new DataType(LocalDate.class));
-	}
-	
+    private static final long serialVersionUID = 1L;
+
+    /**
+     * Constructor.
+     */
+    @SuppressWarnings("unused")
+    public AttributeDate() {
+        super();
+        // dummy for hibernate
+    }
+
+    /**
+     * Constructor.
+     *
+     * @param aCode The unique code for this Question
+     * @param aName The human readable summary name
+     */
+    public AttributeDate(String aCode, String aName) {
+        super(aCode, aName, new DataType(LocalDate.class));
+    }
+
 }
