@@ -21,7 +21,9 @@
 
 package life.genny.qwandaq.attribute;
 
-import java.io.Serializable;
+import com.querydsl.core.annotations.QueryExclude;
+import life.genny.qwandaq.datatype.DataType;
+import life.genny.qwandaq.entity.BaseEntity;
 
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
@@ -30,11 +32,7 @@ import javax.persistence.InheritanceType;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlRootElement;
-
-import life.genny.qwandaq.datatype.DataType;
-import life.genny.qwandaq.entity.BaseEntity;
-
-import com.querydsl.core.annotations.QueryExclude;
+import java.io.Serializable;
 
 /**
  * AttributeLink class is the main link type between BaseEntitys.
@@ -45,13 +43,12 @@ import com.querydsl.core.annotations.QueryExclude;
  * <p>
  * AttributeLink represent the major way of specifying the link between a target
  * from sources.
- * <p>
- * 
- * 
- * @author      Adam Crow
- * @author      Byron Aguirre
- * @version     %I%, %G%
- * @since       1.0
+ * </p>
+ *
+ * @author Adam Crow
+ * @author Byron Aguirre
+ * @version %I%, %G%
+ * @since 1.0
  */
 
 @SuppressWarnings("serial")
@@ -64,38 +61,36 @@ import com.querydsl.core.annotations.QueryExclude;
 @DiscriminatorValue("link")
 public class AttributeLink extends Attribute implements Serializable {
 
-	private static final long serialVersionUID = 1L;
+    private static final long serialVersionUID = 1L;
 
-	private static final String DEFAULT_LNKCODE_PREFIX = "LNK_";
-		
-	/**
-	 * Constructor.
-	 */
-	@SuppressWarnings("unused")
-	public AttributeLink()
-	{
-		super();
-		// dummy for hibernate
-	}
-	
-	/**
-	 * Constructor.
-	 * 
-	 * @param aCode The unique code for this Question
-	 * @param aName The human readable summary name
-	 */
-	public AttributeLink(String aCode, String aName)
-	{
-		super(aCode, aName, new DataType(BaseEntity.class));
-	}
-	
-	/**
-	 * getDefaultCodePrefix This method is overrides the Base class
-	 * 
-	 * @return the default Code prefix for this class.
-	 */
-	static public String getDefaultCodePrefix() { // TODO: this won't work
-		return DEFAULT_LNKCODE_PREFIX;
-	}
-	
+    private static final String DEFAULT_LNKCODE_PREFIX = "LNK_";
+
+    /**
+     * Constructor.
+     */
+    @SuppressWarnings("unused")
+    public AttributeLink() {
+        super();
+        // dummy for hibernate
+    }
+
+    /**
+     * Constructor.
+     *
+     * @param aCode The unique code for this Question
+     * @param aName The human readable summary name
+     */
+    public AttributeLink(String aCode, String aName) {
+        super(aCode, aName, new DataType(BaseEntity.class));
+    }
+
+    /**
+     * getDefaultCodePrefix This method is overrides the Base class
+     *
+     * @return the default Code prefix for this class.
+     */
+    static public String getDefaultCodePrefix() { // TODO: this won't work
+        return DEFAULT_LNKCODE_PREFIX;
+    }
+
 }
