@@ -21,7 +21,8 @@
 
 package life.genny.qwandaq.attribute;
 
-import java.io.Serializable;
+import com.querydsl.core.annotations.QueryExclude;
+import life.genny.qwandaq.datatype.DataType;
 
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
@@ -30,10 +31,7 @@ import javax.persistence.InheritanceType;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlRootElement;
-
-import life.genny.qwandaq.datatype.DataType;
-
-import com.querydsl.core.annotations.QueryExclude;
+import java.io.Serializable;
 
 /**
  * AttributeInteger class handles LocalInteger based attributes.
@@ -44,13 +42,12 @@ import com.querydsl.core.annotations.QueryExclude;
  * <p>
  * AttributeInteger represent the major way of specifying the Integer data type about a target
  * from sources.
- * <p>
- * 
- * 
- * @author      Adam Crow
- * @author      Byron Aguirre
- * @version     %I%, %G%
- * @since       1.0
+ * </p>
+ *
+ * @author Adam Crow
+ * @author Byron Aguirre
+ * @version %I%, %G%
+ * @since 1.0
  */
 
 @SuppressWarnings("serial")
@@ -63,27 +60,25 @@ import com.querydsl.core.annotations.QueryExclude;
 @DiscriminatorValue("long")
 public class AttributeLong extends Attribute implements Serializable {
 
-	private static final long serialVersionUID = 1L;
+    private static final long serialVersionUID = 1L;
 
-	/**
-	 * Constructor.
-	 */
-	@SuppressWarnings("unused")
-	public AttributeLong()
-	{
-		super();
-		// dummy for hibernate
-	}
-	
-	/**
-	 * Constructor.
-	 * 
-	 * @param aCode The unique code for this Question
-	 * @param aName The human readable summary name
-	 */
-	public AttributeLong(String aCode, String aName)
-	{
-		super(aCode, aName, new DataType(Long.class));
-	}
-	
+    /**
+     * Constructor.
+     */
+    @SuppressWarnings("unused")
+    public AttributeLong() {
+        super();
+        // dummy for hibernate
+    }
+
+    /**
+     * Constructor.
+     *
+     * @param aCode The unique code for this Question
+     * @param aName The human readable summary name
+     */
+    public AttributeLong(String aCode, String aName) {
+        super(aCode, aName, new DataType(Long.class));
+    }
+
 }
