@@ -21,7 +21,9 @@
 
 package life.genny.qwandaq.attribute;
 
-import java.io.Serializable;
+import com.querydsl.core.annotations.QueryExclude;
+import life.genny.qwandaq.datatype.DataType;
+import org.javamoney.moneta.Money;
 
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
@@ -30,12 +32,7 @@ import javax.persistence.InheritanceType;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlRootElement;
-
-import org.javamoney.moneta.Money;
-
-import life.genny.qwandaq.datatype.DataType;
-
-import com.querydsl.core.annotations.QueryExclude;
+import java.io.Serializable;
 
 /**
  * AttributeMoney class handles LocalDate based attributes.
@@ -46,13 +43,12 @@ import com.querydsl.core.annotations.QueryExclude;
  * <p>
  * AttributeDate represent the major way of specifying the Money data type about a target
  * from sources.
- * <p>
- * 
- * 
- * @author      Adam Crow
- * @author      Byron Aguirre
- * @version     %I%, %G%
- * @since       1.0
+ * </p>
+ *
+ * @author Adam Crow
+ * @author Byron Aguirre
+ * @version %I%, %G%
+ * @since 1.0
  */
 
 @SuppressWarnings("serial")
@@ -65,28 +61,26 @@ import com.querydsl.core.annotations.QueryExclude;
 @DiscriminatorValue("money")
 public class AttributeMoney extends Attribute implements Serializable {
 
-	private static final long serialVersionUID = 1L;
-	
-	/**
-	 * Constructor.
-	 */
-	@SuppressWarnings("unused")
-	public AttributeMoney()
-	{
-		super();
-		// dummy for hibernate
-	}
-	
-	/**
-	 * Constructor.
-	 * 
-	 * @param aCode The unique code for this Question
-	 * @param aName The human readable summary name
-	 */
-	public AttributeMoney(String aCode, String aName)
-	{
-		super(aCode, aName, new DataType(Money.class));
-		
-	}
-	
+    private static final long serialVersionUID = 1L;
+
+    /**
+     * Constructor.
+     */
+    @SuppressWarnings("unused")
+    public AttributeMoney() {
+        super();
+        // dummy for hibernate
+    }
+
+    /**
+     * Constructor.
+     *
+     * @param aCode The unique code for this Question
+     * @param aName The human readable summary name
+     */
+    public AttributeMoney(String aCode, String aName) {
+        super(aCode, aName, new DataType(Money.class));
+
+    }
+
 }
