@@ -1,12 +1,11 @@
 package life.genny.qwandaq.converter;
 
 
+import javax.persistence.AttributeConverter;
+import javax.persistence.Converter;
 import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
-
-import javax.persistence.AttributeConverter;
-import javax.persistence.Converter;
 
 @Converter
 public class StringListConverter implements AttributeConverter<List<String>, String> {
@@ -39,7 +38,7 @@ public class StringListConverter implements AttributeConverter<List<String>, Str
    */
   @Override
   public List<String> convertToEntityAttribute(final String joined) {
-    List<String> strings = new CopyOnWriteArrayList<String>();
+      List<String> strings = new CopyOnWriteArrayList<>();
     if (joined != null) {
       strings = new CopyOnWriteArrayList<>(Arrays.asList(joined.split(",")));
     }
