@@ -1,18 +1,18 @@
 package life.genny.qwandaq.serialization.baseentity;
 
-import java.io.IOException;
-
 import org.infinispan.protostream.MessageMarshaller;
+
+import java.io.IOException;
 
 public class BaseEntityKeyMessageMarshaller implements MessageMarshaller<BaseEntityKey> {
 
-	@Override
-	public Class<BaseEntityKey> getJavaClass() {
-		return BaseEntityKey.class;
-	}
+    @Override
+    public Class<BaseEntityKey> getJavaClass() {
+        return BaseEntityKey.class;
+    }
 
-	@Override
-	public String getTypeName() {
+    @Override
+    public String getTypeName() {
 		return "life.genny.qwandaq.serialization.baseentity.BaseEntityKey";
 	}
 
@@ -20,8 +20,7 @@ public class BaseEntityKeyMessageMarshaller implements MessageMarshaller<BaseEnt
 	public BaseEntityKey readFrom(ProtoStreamReader reader) throws IOException {
 		String productCode = reader.readString("realm");
 		String beCode = reader.readString("code");
-		BaseEntityKey bek = new BaseEntityKey(productCode, beCode);
-		return bek;
+        return new BaseEntityKey(productCode, beCode);
 	}
 
 	@Override
