@@ -1,23 +1,17 @@
 package life.genny.qwandaq.utils;
 
-import java.io.StringReader;
-import java.text.DecimalFormat;
-import java.time.LocalDateTime;
-import java.time.LocalDate;
-import java.util.Map;
-import java.util.HashMap;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import life.genny.qwandaq.entity.BaseEntity;
+import org.jboss.logging.Logger;
 
 import javax.json.Json;
 import javax.json.JsonArray;
 import javax.json.JsonObject;
 import javax.json.JsonReader;
-
-import org.jboss.logging.Logger;
-import org.javamoney.moneta.Money;
-
-import life.genny.qwandaq.entity.BaseEntity;
+import java.io.StringReader;
+import java.text.DecimalFormat;
+import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 /**
  * A utiltity used in the MailMerge feature of Genny.
@@ -108,7 +102,7 @@ public class MergeUtils {
 	public static Object wordMerge(String mergeText, Map<String, Object> entitymap) {
 
 		if (mergeText == null || mergeText.isEmpty()) {
-			return DEFAULT;	
+			return DEFAULT;
 		}
 
 		// we split the text to merge into 2 components: BE.PRI... becomes [BE, PRI...]
@@ -116,8 +110,8 @@ public class MergeUtils {
 		String keyCode = entityArr[0];
 		log.debug("looking for key in map: " + keyCode);
 
-		if ((entityArr.length == 0))
-			return DEFAULT;
+//		if ((entityArr.length == 0))
+//			return DEFAULT;
 
 		if (!entitymap.containsKey(keyCode)) {
 			return DEFAULT;
@@ -144,19 +138,18 @@ public class MergeUtils {
 			log.info("context: " + keyCode + ", attr: " + attributeCode + ", value: " + attributeValue);
 
 			Matcher matchFormat = null;
-			if (entityArr != null && entityArr.length > 2) {
+			if (entityArr.length > 2) {
 				matchFormat = FORMAT_PATTERN_VARIABLE.matcher(entityArr[2]);
 			}
 
-			if (attributeValue instanceof org.javamoney.moneta.Money) {
+			if (attributeValue instanceof org.javamoney.moneta.Money money) {
 
 				log.info("This is a Money attribute");
-				DecimalFormat df = new DecimalFormat("#.00"); 
-				Money money = (Money) attributeValue; 
+				DecimalFormat df = new DecimalFormat("#.00");
 
 				return df.format(money.getNumber()) + " " + money.getCurrency();
 
-			} else if (attributeValue instanceof java.time.LocalDateTime) {
+			} else if (attributeValue instanceof java.time.LocalDateTime ldtValue) {
 				/*
 				   If the date-related mergeString needs to format to a particultar 
 				   format -> we split the date-time related merge text to merge 
@@ -167,27 +160,27 @@ public class MergeUtils {
 				   */
 				if (matchFormat != null && matchFormat.find()) {
 					log.info("Datetime attribute " + attributeCode + " needs formatting. Format is " + entityArr[2]);
-					return TimeUtils.formatDateTime((LocalDateTime) attributeValue, matchFormat.group(1));
+					return TimeUtils.formatDateTime(ldtValue, matchFormat.group(1));
 				} else {
 					log.info("DateTime attribute " + attributeCode + " does NOT need formatting");
-					return (LocalDateTime) attributeValue;
+					return ldtValue;
 				}
 
-			} else if (attributeValue instanceof java.time.LocalDate) {
+			} else if (attributeValue instanceof java.time.LocalDate ldValue) {
 
 				if (matchFormat != null && matchFormat.find()) {
 					log.info("Date attribute " + attributeCode + " needs formatting. Format is " + entityArr[2]);
-					return TimeUtils.formatDate((LocalDate) attributeValue, matchFormat.group(1));
+					return TimeUtils.formatDate(ldValue, matchFormat.group(1));
 				} else {
 					log.info("Date attribute " + attributeCode + " does NOT need formatting");
-					return (LocalDate) attributeValue;
+					return ldValue;
 				}
 
-			} else if (attributeValue instanceof java.lang.String) {
+			} else if (attributeValue instanceof java.lang.String sValue) {
 
 				String result = null;
 				if (matchFormat != null && matchFormat.find()) {
-					result  =  getFormattedString((String) attributeValue, matchFormat.group(1));
+					result = getFormattedString(sValue, matchFormat.group(1));
 					log.info("String attribute " + attributeCode + " needs formatting. Format is " + entityArr[2] + ", Result is " + result);
 				} else {
 					result = be.findEntityAttribute(attributeCode).get().getValueString();
@@ -195,20 +188,20 @@ public class MergeUtils {
 				}
 				return result;
 
-			} else if (attributeValue instanceof java.lang.Boolean) {
-				return (Boolean) attributeValue;
-			} else if (attributeValue instanceof java.lang.Integer) {
-				return (Integer) attributeValue;
-			} else if (attributeValue instanceof java.lang.Long) {
-				return (Long) attributeValue;
-			} else if (attributeValue instanceof java.lang.Double) {
-				return (Double) attributeValue;
+			} else if (attributeValue instanceof java.lang.Boolean bValue) {
+				return bValue;
+			} else if (attributeValue instanceof java.lang.Integer iValue) {
+				return iValue;
+			} else if (attributeValue instanceof java.lang.Long lValue) {
+				return lValue;
+			} else if (attributeValue instanceof java.lang.Double dValue) {
+				return dValue;
 			} else {
 				return be.findEntityAttribute(attributeCode).get().getValueString();
 			}
 
 		} else if (value.getClass().equals(String.class)) {
-			return (String) value;
+			return value;
 		}
 
 		return DEFAULT;	
